+++ CHANGELOG.md
@@ -79,6 +79,7 @@
 
 ### Bugs fixed
 
+- [#1445](https://github.com/bbatsov/prelude/issues/1445): prelude-rust fails to load due to tree-sitter dependency issue
 - [#1302](https://github.com/bbatsov/prelude/issues/1302): `C-a` should be bound to `org-beginning-of-line` in org-mode buffers.
 
 ## 1.0.0 (2020-09-15)